[dashboard] clean-up code
